+++ src/app/client/src/app/modules/workspace/components/assessments/assessments-list/assessments-list.component.html
@@ -1,0 +1,66 @@
+<!-- <app-workspace-content-filter ></app-workspace-content-filter> -->
+<div class="twelve wide column mt-20" *ngIf="showLoader">
+    <app-loader [data]="loaderMessage"></app-loader>
+</div>
+<div *ngIf="noResult && !showLoader">
+    <app-no-result [data]="noResultMessage"></app-no-result>
+</div>
+
+<table class="ui upForReviewTable table" *ngIf="allAssessments.length > 0 && !showLoader && !noResult ">
+    <thead>
+        <tr>
+            <th class="four wide UpReviewTableHeader">{{ resourceService?.frmelmnts?.snav?.assessments }}</th>
+            <th class="seven wide UpReviewTableHeader">{{ resourceService?.frmelmnts?.lbl?.header?.position }}</th>
+            <th class="three wide UpReviewTableHeader">{{ resourceService?.frmelmnts?.lbl?.action }}</th>
+        </tr>
+    </thead>
+    <tbody [appTelemetryImpression]="telemetryImpression" in-view-container (inview)="inview($event)"
+        [throttle]="[1000]" [trigger]="allAssessments">
+        <tr in-view-item class="UpForReviewListHover" *ngFor="let assessment of allAssessments let i = index;" [id]="i"
+            [data]="content" appTelemetryInteract>
+            <td tabindex="0">
+                <h5 class="ui image header">
+                    <div class="content ">
+                        <div class="UpReviewHeader">{{ (assessment?.name) ? (assessment?.name | slice:0:50) : 'NA' }}
+                            <i *ngIf="assessment?.name?.length > 50">...</i>
+                        </div>
+                    </div>
+                </h5>
+            </td>
+            <td>
+                <h5 class="ui image header">
+                    <div class="content">
+                        <div class="UpReviewSubHeader">General Nursing and Midwifery</div>
+                        <hr>
+                        <div class="UpReviewSubHeader">C1, L4 &nbsp;&nbsp; C1, L4 &nbsp;&nbsp; C1, L4</div>
+                        <hr>
+                        <div class="UpReviewSubHeader">C1, L4</div>
+                    </div>
+                </h5>
+            </td>
+            <td>
+                <div class="assign-link" [routerLink]="['/workspace/content/assessments/assign']">Assign</div>
+            </td>
+        </tr>
+    </tbody>
+</table>
+<div class="ui grid">
+    <div class="two wide column"></div>
+    <div class="ten wide column">
+        <div class="pb-10"
+            *ngIf="allContent && totalCount > config.appConfig.WORKSPACE.PAGE_LIMIT && !showLoader && !showError && !noResult">
+            <div class="ui pagination menu mt-10 right-floated p-0 " *ngIf="pager.pages.length ">
+                <a [ngClass]="{disabled:pager.currentPage===1 }" class="item " tabindex="0"
+                    (click)="navigateToPage(1) ">First</a>
+                <a [ngClass]="{disabled:pager.currentPage===1 }" class="item " tabindex="0"
+                    (click)="navigateToPage(pager.currentPage - 1)">Previous</a>
+                <a *ngFor="let page of pager.pages" [ngClass]="{active:pager.currentPage===page}" tabindex="0"
+                    (click)="navigateToPage(page)" class="item">{{page}}</a>
+                <a [ngClass]="{disabled:pager.currentPage=== pager.totalPages}" tabindex="0"
+                    (click)="navigateToPage(pager.currentPage + 1)" class="item">Next</a>
+                <a [ngClass]="{disabled:pager.currentPage=== pager.totalPages}" tabindex="0"
+                    (click)="navigateToPage(pager.totalPages)" class="item ">Last</a>
+            </div>
+        </div>
+    </div>
+</div>
