+++ src/app/client/src/app/modules/workspace/components/assessments/assessments-list/assessments-list.component.scss
