+++ src/app/client/src/app/modules/workspace/components/assessments/students-list/students-list.component.html
@@ -3,7 +3,7 @@
 </div>
 
 <div *ngIf="!showLoader">
-    <button class="sb-btn sb-btn-primary sb-btn-normal sb-btn-border" routerLink="/workspace/content/assessments">
+    <button class="sb-btn sb-btn-primary sb-btn-normal sb-btn-border" routerLink="/workspace/content/assessments/list">
         <i aria-hidden="true" class="arrow left icon"></i>
         Back to assignments
     </button>
